New class for retry section
