+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/commons/RetrySection.composable.kt
@@ -1,0 +1,30 @@
+package com.example.zeldacompendium.presentation.ui.commons
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun RetrySection(
+   error: String,
+   onRetry: () -> Unit
+) {
+   Column {
+      Text(error, color = Color.Red, fontSize = 18.sp)
+      Spacer(modifier = Modifier.height(8.dp))
+      Button(
+         onClick = { onRetry() },
+         modifier = Modifier.align(Alignment.CenterHorizontally)
+      ) {
+         Text(text = "Retry")
+      }
+   }
+}
