+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/lists/breath/components/List.composable.kt
@@ -1,27 +1,21 @@
 package com.example.zeldacompendium.presentation.ui.lists.breath.components
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.zeldacompendium.data.models.CompendiumListEntry
+import com.example.zeldacompendium.presentation.ui.commons.RetrySection
 import com.example.zeldacompendium.presentation.ui.lists.ImageList
 import com.example.zeldacompendium.presentation.ui.lists.breath.CompendiumBreathViewModel
 
@@ -60,23 +54,6 @@ fun CompendiumList(
          RetrySection(error = loadError) {
             viewModel.loadBreathList()
          }
-      }
-   }
-}
-
-@Composable
-fun RetrySection(
-   error: String,
-   onRetry: () -> Unit
-) {
-   Column {
-      Text(error, color = Color.Red, fontSize = 18.sp)
-      Spacer(modifier = Modifier.height(8.dp))
-      Button(
-         onClick = { onRetry() },
-         modifier = Modifier.align(Alignment.CenterHorizontally)
-      ) {
-         Text(text = "Retry")
       }
    }
 }
