+++ app/src/main/java/com/example/zeldacompendium/presentation/ui/lists/tears/TearsContainer.composable.kt
@@ -48,6 +48,7 @@ fun TearsContainer(
    var selectedIndex by remember { mutableIntStateOf(0) }
    var searchText by remember { mutableStateOf("") }
    val isLoading by remember { viewModel.isLoading }
+   val isError by remember { viewModel.loadError }
    val filteredList = compendiumFilter.filterCompendiumList(viewModel.compendiumList.value, selectedIndex)
 
    Scaffold(
@@ -93,7 +94,7 @@ fun TearsContainer(
             .padding(padding)
             .fillMaxSize()
       ) {
-         if (filteredList.isEmpty() && !isLoading) {
+         if (filteredList.isEmpty() && !isLoading && isError.isEmpty()) {
             Column(modifier = Modifier.padding(16.dp)) {
                ImageList(gameId = gameId)
                CompendiumItemBreathEmpty()
